# server/openapi.yaml
openapi: 3.0.3
info:
  title: Pense Aja Backend API
  version: 1.0.0
  description: >
    API backend do Pense Aja para cadastro e avaliação de ideias, pontuação,
    marketplace de recompensas, perfil de usuário, RBAC, dashboards e apoio de IA.
servers:
  - url: http://localhost:2512
    description: Desenvolvimento local
...
security:
  - cookieAuth: []
paths:
...
  /marketplace/catalog/{dassOffice}:
    get:
      tags: [Marketplace]
      summary: Listar catálogo operacional
      description: >
        Lista itens do catálogo operacional da unidade Dass. Acesso público no
        código atual.
      security: []
      parameters:
        - $ref: "#/components/parameters/DassOfficePath"
      responses:
        "200":
          description: Catálogo encontrado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceCatalogItem"
              example:
                - id: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
                  legacyProductId: "12"
                  dassOffice: "SEST"
                  name: "Garrafa térmica"
                  imageUrl: "https://cdn.exemplo.com/garrafa.png"
                  pointsCost: 150
                  itemType: "physical"
                  active: true
                  availableQuantity: null
                  metadata:
                    source: "pense_aja_loja"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags: [Marketplace]
      summary: Criar ou atualizar itens de catálogo
      description: >
        Cria ou atualiza itens do catálogo operacional. Exige token válido e
        permissão `catalog.manage` na unidade.
      parameters:
        - $ref: "#/components/parameters/DassOfficePath"
        - $ref: "#/components/parameters/CookieHeader"
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UpsertCatalogItemInput"
            example:
              - id: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
                name: "Garrafa térmica Dass"
                imageUrl: "https://cdn.exemplo.com/garrafa.png"
                pointsCost: 180
                itemType: "physical"
                active: true
                availableQuantity: 25
                metadata:
                  supplier: "almoxarifado"
      responses:
        "200":
          description: Catálogo atualizado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceCatalogItem"
              example:
                - id: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
                  legacyProductId: "12"
                  dassOffice: "SEST"
                  name: "Garrafa térmica Dass"
                  imageUrl: "https://cdn.exemplo.com/garrafa.png"
                  pointsCost: 180
                  itemType: "physical"
                  active: true
                  availableQuantity: 25
                  metadata:
                    supplier: "almoxarifado"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /marketplace/requests:
    post:
      tags: [Marketplace]
      summary: Solicitar resgate no marketplace
      description: >
        Cria solicitação de resgate, reserva pontos no ledger e deixa a
        solicitação em `pending_approval`. Exige token válido e permissão
        `marketplace.request.create` na unidade.
      parameters:
        - $ref: "#/components/parameters/CookieHeader"
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketplaceRequestInput"
            example:
              dassOffice: "SEST"
              catalogItemId: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
              registration: 1234567
              requesterName: "Maria Souza"
              reason: "Resgate de premiação por pontos acumulados."
      responses:
        "201":
          description: Solicitação criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceRequest"
              example:
                id: 927
                registration: "1234567"
                dassOffice: "SEST"
                catalogItemId: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
                requestStatus: "pending_approval"
                reservedLedgerEntryId: 451
                approvalActorRegistration: null
                approvalActorName: null
                fulfillmentType: "none"
                legacyPrizeId: null
                createdAt: "2026-04-22T12:00:00.000Z"
                updatedAt: "2026-04-22T12:00:00.000Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Marketplace]
      summary: Listar solicitações de marketplace
      description: >
        Lista solicitações da unidade para operação de aprovação/fulfillment.
        Exige token válido e permissão `marketplace.request.approve`.
      parameters:
        - $ref: "#/components/parameters/DassOfficeQueryRequired"
        - $ref: "#/components/parameters/CookieHeader"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Solicitações encontradas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceRequest"
              example:
                - id: 927
                  registration: "1234567"
                  dassOffice: "SEST"
                  catalogItemId: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
                  requestStatus: "pending_approval"
                  reservedLedgerEntryId: 451
                  approvalActorRegistration: null
                  approvalActorName: null
                  fulfillmentType: "none"
                  legacyPrizeId: null
                  createdAt: "2026-04-22T12:00:00.000Z"
                  updatedAt: "2026-04-22T12:00:00.000Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /marketplace/requests/{id}/approve:
    put:
      tags: [Marketplace]
      summary: Aprovar solicitação de resgate
      description: >
        Aprova uma solicitação pendente, confirma o consumo dos pontos com
        lançamento `commit` e finaliza o resgate como `completed`. Não gera
        voucher ou fulfillment. Exige token válido e permissão
        `marketplace.request.approve` na unidade enviada no payload.
      parameters:
        - $ref: "#/components/parameters/RequestIdPath"
        - $ref: "#/components/parameters/CookieHeader"
      security:
        - cookieAuth: []
      requestBody:
        $ref: "#/components/requestBodies/MarketplaceTransition"
      responses:
        "200":
          $ref: "#/components/responses/MarketplaceRequestApproved"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /marketplace/requests/{id}/reject:
    put:
      tags: [Marketplace]
      summary: Rejeitar solicitação de resgate
      description: >
        Rejeita uma solicitação e libera pontos reservados quando existir reserva.
        Exige token válido e permissão `marketplace.request.approve`.
      parameters:
        - $ref: "#/components/parameters/RequestIdPath"
        - $ref: "#/components/parameters/CookieHeader"
      security:
        - cookieAuth: []
      requestBody:
        $ref: "#/components/requestBodies/MarketplaceTransition"
      responses:
        "200":
          $ref: "#/components/responses/MarketplaceRequestRejected"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /marketplace/requests/{id}/refund:
    put:
      tags: [Marketplace]
      summary: Estornar solicitação concluída
      description: >
        Estorna uma solicitação concluída com lançamento `refund` no ledger.
        Exige token válido e permissão `marketplace.request.approve` ou
        `marketplace.refund`.
      parameters:
        - $ref: "#/components/parameters/RequestIdPath"
        - $ref: "#/components/parameters/CookieHeader"
      security:
        - cookieAuth: []
      requestBody:
        $ref: "#/components/requestBodies/MarketplaceTransition"
      responses:
        "200":
          $ref: "#/components/responses/MarketplaceRequestRefunded"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Token JWT de sessão enviado no cookie `token`.
  parameters:
    CookieHeader:
      name: Cookie
      in: header
      required: true
      schema:
        type: string
      description: Cookie de autenticação contendo `token=<jwt>`.
      example: "token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    DassOfficePath:
      name: dassOffice
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/DassOffice"
      description: Unidade Dass da operação.
      example: "SEST"
    DassOfficeQueryRequired:
      name: dassOffice
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/DassOffice"
      description: Unidade Dass da consulta.
      example: "SEST"
...
    RequestIdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Identificador da solicitação de marketplace.
      example: 927
...
  requestBodies:
    MarketplaceTransition:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MarketplaceTransitionInput"
          example:
            dassOffice: "SEST"
            reason: "Solicitação validada pela liderança."
            notes: "Sem pendências operacionais."
            metadata:
              channel: "admin_panel"
  responses:
    ValidationError:
      description: Erro de validação de entrada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            message: "Dados inválidos. Unidade Dass inválida!"
...
    Unauthorized:
      description: Usuário não autenticado ou token ausente/inválido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            message: "Usuário não autenticado."
    Forbidden:
      description: Usuário autenticado sem permissão para a operação.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            message: "Acesso negado. Permissão insuficiente."
    InternalServerError:
      description: Erro interno tratado pelo middleware global.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            message: "Erro interno no servidor. - Contate a equipe de automação!"
    MarketplaceRequestApproved:
      description: Solicitação aprovada e concluída com desconto de pontos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MarketplaceRequest"
          example:
            id: 927
            registration: "1234567"
            dassOffice: "SEST"
            catalogItemId: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
            requestStatus: "completed"
            reservedLedgerEntryId: 451
            approvalActorRegistration: "7654321"
            approvalActorName: "João Analista"
            fulfillmentType: "none"
            legacyPrizeId: null
            createdAt: "2026-04-22T12:00:00.000Z"
            updatedAt: "2026-04-22T13:00:00.000Z"
    MarketplaceRequestRejected:
      description: Solicitação rejeitada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MarketplaceRequest"
          example:
            id: 927
            registration: "1234567"
            dassOffice: "SEST"
            catalogItemId: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
            requestStatus: "rejected"
            reservedLedgerEntryId: null
            approvalActorRegistration: "7654321"
            approvalActorName: "João Analista"
            fulfillmentType: "none"
            legacyPrizeId: null
            createdAt: "2026-04-22T12:00:00.000Z"
            updatedAt: "2026-04-22T13:00:00.000Z"
    MarketplaceRequestRefunded:
      description: Solicitação estornada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MarketplaceRequest"
          example:
            id: 927
            registration: "1234567"
            dassOffice: "SEST"
            catalogItemId: "6f60f952-ea22-49b9-a1dd-f346f42f38f2"
            requestStatus: "refunded"
            reservedLedgerEntryId: null
            approvalActorRegistration: "7654321"
            approvalActorName: "João Analista"
            fulfillmentType: "physical_delivery"
            legacyPrizeId: null
            createdAt: "2026-04-22T12:00:00.000Z"
            updatedAt: "2026-04-23T12:00:00.000Z"
  schemas:
    DassOffice:
      type: string
      enum: [SEST, VDC, ITB, VDC-CONF, STJ]
    ErrorMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    MarketplaceItemType:
      type: string
      enum: [physical, voucher]
    MarketplaceRequestStatus:
      type: string
      enum:
        - pending_approval
        - approved
        - rejected
        - fulfillment_in_progress
        - completed
        - cancelled
        - refunded
    MarketplaceCatalogItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legacyProductId:
          type: string
          nullable: true
        dassOffice:
          $ref: "#/components/schemas/DassOffice"
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        pointsCost:
          type: integer
        itemType:
          $ref: "#/components/schemas/MarketplaceItemType"
        active:
          type: boolean
        availableQuantity:
          type: integer
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    UpsertCatalogItemInput:
      type: object
      required: [name, pointsCost]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        pointsCost:
          type: integer
        itemType:
          $ref: "#/components/schemas/MarketplaceItemType"
        active:
          type: boolean
        availableQuantity:
          type: integer
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    CreateMarketplaceRequestInput:
      type: object
      required: [dassOffice, catalogItemId]
      properties:
        dassOffice:
          $ref: "#/components/schemas/DassOffice"
        catalogItemId:
          type: string
          format: uuid
        registration:
          oneOf:
            - type: string
            - type: integer
        requesterName:
          type: string
        reason:
          type: string
    MarketplaceTransitionInput:
      type: object
      required: [dassOffice]
      properties:
        dassOffice:
          $ref: "#/components/schemas/DassOffice"
        reason:
          type: string
        notes:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    MarketplaceRequest:
      type: object
      properties:
        id:
          type: integer
        registration:
          type: string
        dassOffice:
          $ref: "#/components/schemas/DassOffice"
        catalogItemId:
          type: string
          format: uuid
        requestStatus:
          $ref: "#/components/schemas/MarketplaceRequestStatus"
        reservedLedgerEntryId:
          type: integer
          nullable: true
        approvalActorRegistration:
          type: string
          nullable: true
        approvalActorName:
          type: string
          nullable: true
        fulfillmentType:
          type: string
        legacyPrizeId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...